+++ packages/kysely/migrate/migrations/index.ts
@@ -86,6 +86,7 @@ import * as n86_index_sessions_login_session_id from "./2025-06-04T13:00:00_inde
 import * as n87_code_challenge from "./2025-06-04T14:00:00_code_challenge";
 import * as n88_add_redirect_uri_to_codes from "./2025-06-05T09:40:00_add_redirect_uri_to_codes";
 import * as n89_add_nonce_and_state_to_codes from "./2025-06-05T14:30:00_add_nonce_and_state_to_codes";
+import * as n90_themes from "./2025-07-23T14:30:00_themes";
 
 // These need to be in alphabetic order
 export default {
@@ -177,4 +178,5 @@ export default {
   n87_code_challenge,
   n88_add_redirect_uri_to_codes,
   n89_add_nonce_and_state_to_codes,
+  n90_themes,
 };