+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/demo
 
+## 0.14.17
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.176.0
+  - @authhero/kysely-adapter@10.24.0
+
 ## 0.14.16
 
 ### Patch Changes